+++ monkey/infection_monkey/payload/ransomware/consts.py
@@ -1,4 +1,0 @@
-from pathlib import Path
-
-README_SRC = Path(__file__).parent / "ransomware_readme.txt"
-README_FILE_NAME = "README.txt"