+++ monkey/monkey_island/cc/ui/src/services/configuration/payloads.js
@@ -1,0 +1,7 @@
+const PAYLOADS = {
+    'title': 'Payloads',
+    'description': 'Click on a payload for more information.',
+    'type': 'object',
+    'properties':{}
+}
+export default PAYLOADS